Finished csv ingestion and processing code

+++ db/migrate/20240725204618_create_records_of_work.rb
@@ -1,7 +1,7 @@
 class CreateRecordsOfWork < ActiveRecord::Migration[7.1]
   def change
     create_table :records_of_work do |t|
-      t.integer :hours
+      t.float :hours
       t.date :date
       t.date :pay_period_start
       t.date :pay_period_end
